chore: -[x] move cmd/indexer/internal/datastore to cmd/internal/datastore -[x]  update all references to this new package location #478
