+++ cmd/internal/datastore/public_key.go
